+++ packages/elements/user-picker/example-helpers/index.ts
@@ -1,0 +1,4 @@
+import { userPickerData } from '@atlaskit/util-data-test';
+import { User } from '../src/types';
+
+export const exampleUsers = userPickerData as User[];