+++ packages/media/media-card/src/utils/getDataURIFromFileState.ts
@@ -4,7 +4,11 @@ import { FileState, getMediaTypeFromMimeType } from '@atlaskit/media-core';
 export const getDataURIFromFileState = async (
   state: FileState,
 ): Promise<string | undefined> => {
-  if (state.status === 'error' || !state.preview) {
+  if (
+    state.status === 'error' ||
+    state.status === 'failed-processing' ||
+    !state.preview
+  ) {
     return undefined;
   }
   const type = state.preview.blob.type;