+++ packages/media/media-viewer/__tests__/newgen/viewers/doc/index.spec.tsx
@@ -31,7 +31,6 @@ const item: ProcessedFileState = {
   size: 11222,
   mediaType: 'video',
   mimeType: 'mp4',
-  binaryUrl: '',
   artifacts: {
     'document.pdf': {
       url: '/pdf',